pyproject.toml
[tool.poetry]
name = "grpc_demo"
version = "0.1.0"
description = ""
authors = ["Anton <Kondrahin210@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.51.3"
grpcio-tools = "^1.51.3"
redis = "^4.5.1"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

